+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Product Backlog
 * [High] Build ShoppingCart, ProductList, and User using Vuex
-* [High] Create AddToCalendar component
 * [High] Create GoogleButton component
 * [High] Create UserManager component for Admin and SHYnet views
 * [High] Create components for Admin tabs (Dashboard, UserManager, ScheduleManager, AnnouncementManager, OrderManager, ProductManager, ClassManager, WorkshopManager, LocationManager)
+* [High] Change modals to b-dropdown-form
 * [Medium] Create JSONLd component
 
 <a name="1.0.0"></a>
+* Create AddToCalendar component
 * Created Tweet component or use 3rd party